# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MilitaryMetrics {
  /** @nullable */
  activePersonnel?: number | null;
  /** @nullable */
  reservePersonnel?: number | null;
  /** @nullable */
  paramilitary?: number | null;
  /** @nullable */
  defenseBudgetUsd?: number | null;
  /** @nullable */
  gdpUsd?: number | null;
  /** @nullable */
  population?: number | null;
  /** @nullable */
  landAreaKm2?: number | null;
  /** @nullable */
  aircraft?: number | null;
  /** @nullable */
  fighterJets?: number | null;
  /** @nullable */
  attackHelicopters?: number | null;
  /** @nullable */
  transportHelicopters?: number | null;
  /** @nullable */
  tanks?: number | null;
  /** @nullable */
  armoredVehicles?: number | null;
  /** @nullable */
  selfPropelledArtillery?: number | null;
  /** @nullable */
  rocketProjectors?: number | null;
  /** @nullable */
  navalVessels?: number | null;
  /** @nullable */
  submarines?: number | null;
  /** @nullable */
  destroyers?: number | null;
  /** @nullable */
  frigates?: number | null;
  /** @nullable */
  aircraftCarriers?: number | null;
  /** @nullable */
  nuclearWarheads?: number | null;
}

export type DataSourceConfidence =
  (typeof DataSourceConfidence)[keyof typeof DataSourceConfidence];

export const DataSourceConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface DataSource {
  metric: string;
  source: string;
  confidence: DataSourceConfidence;
  lastUpdated: string;
}

export interface Country {
  code: string;
  name: string;
  flagEmoji: string;
  continent: string;
  region: string;
  alliances: string[];
  metrics: MilitaryMetrics;
  dataSources: DataSource[];
}

export interface CategoryWeights {
  manpower: number;
  airPower: number;
  groundForces: number;
  navalForces: number;
  economy: number;
  logistics: number;
}

export type ComparisonRequestScenario =
  (typeof ComparisonRequestScenario)[keyof typeof ComparisonRequestScenario];

export const ComparisonRequestScenario = {
  full: "full",
  air: "air",
  naval: "naval",
  ground: "ground",
} as const;

export interface ComparisonRequest {
  /** @minItems 2 */
  countryCodes: string[];
  weights: CategoryWeights;
  scenario: ComparisonRequestScenario;
}

export type CategoryScoreScores = { [key: string]: number };

export interface CategoryScore {
  category: string;
  scores: CategoryScoreScores;
  advantage: string;
  advantageCountryCode: string;
}

export type CountryScoreCategoryScores = { [key: string]: number };

export interface CountryScore {
  countryCode: string;
  totalScore: number;
  categoryScores: CountryScoreCategoryScores;
}

export interface ComparisonResult {
  countries: Country[];
  scores: CountryScore[];
  categoryBreakdown: CategoryScore[];
  overallWinner: string;
}

export type CoalitionRequestScenario =
  (typeof CoalitionRequestScenario)[keyof typeof CoalitionRequestScenario];

export const CoalitionRequestScenario = {
  full: "full",
  air: "air",
  naval: "naval",
  ground: "ground",
} as const;

export interface CoalitionRequest {
  /** @minItems 1 */
  teamA: string[];
  /** @minItems 1 */
  teamB: string[];
  weights: CategoryWeights;
  scenario: CoalitionRequestScenario;
}

export type CoalitionMetricsCategoryScores = { [key: string]: number };

export interface CoalitionMetrics {
  teamName: string;
  countries: Country[];
  combinedMetrics: MilitaryMetrics;
  totalScore: number;
  categoryScores: CoalitionMetricsCategoryScores;
}

export interface CoalitionResult {
  teamA: CoalitionMetrics;
  teamB: CoalitionMetrics;
  categoryBreakdown: CategoryScore[];
  winner: string;
}

export interface RegionsMetadata {
  continents: string[];
  regions: string[];
}

export interface TimelineEntry {
  year: number;
  /** @nullable */
  defenseBudgetUsd?: number | null;
  /** @nullable */
  activePersonnel?: number | null;
  /** @nullable */
  gdpUsd?: number | null;
}

export type ListCountriesParams = {
  continent?: string;
  alliance?: string;
  region?: string;
};