# lib/api-zod/src/generated/types/categoryScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryScoreScores } from "./categoryScoreScores";

export interface CategoryScore {
  category: string;
  scores: CategoryScoreScores;
  advantage: string;
  advantageCountryCode: string;
}