# lib/api-zod/src/generated/types/coalitionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryWeights } from "./categoryWeights";
import type { CoalitionRequestScenario } from "./coalitionRequestScenario";

export interface CoalitionRequest {
  /** @minItems 1 */
  teamA: string[];
  /** @minItems 1 */
  teamB: string[];
  weights: CategoryWeights;
  scenario: CoalitionRequestScenario;
}