# lib/api-zod/src/generated/types/coalitionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryScore } from "./categoryScore";
import type { CoalitionMetrics } from "./coalitionMetrics";

export interface CoalitionResult {
  teamA: CoalitionMetrics;
  teamB: CoalitionMetrics;
  categoryBreakdown: CategoryScore[];
  winner: string;
}