# lib/api-zod/src/generated/types/comparisonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryWeights } from "./categoryWeights";
import type { ComparisonRequestScenario } from "./comparisonRequestScenario";

export interface ComparisonRequest {
  /** @minItems 2 */
  countryCodes: string[];
  weights: CategoryWeights;
  scenario: ComparisonRequestScenario;
}