# lib/api-zod/src/generated/types/country.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */
import type { DataSource } from "./dataSource";
import type { MilitaryMetrics } from "./militaryMetrics";

export interface Country {
  code: string;
  name: string;
  flagEmoji: string;
  continent: string;
  region: string;
  alliances: string[];
  metrics: MilitaryMetrics;
  dataSources: DataSource[];
}