# lib/api-zod/src/generated/types/countryScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */
import type { CountryScoreCategoryScores } from "./countryScoreCategoryScores";

export interface CountryScore {
  countryCode: string;
  totalScore: number;
  categoryScores: CountryScoreCategoryScores;
}