# lib/api-zod/src/generated/types/militaryMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */

export interface MilitaryMetrics {
  /** @nullable */
  activePersonnel?: number | null;
  /** @nullable */
  reservePersonnel?: number | null;
  /** @nullable */
  paramilitary?: number | null;
  /** @nullable */
  defenseBudgetUsd?: number | null;
  /** @nullable */
  gdpUsd?: number | null;
  /** @nullable */
  population?: number | null;
  /** @nullable */
  landAreaKm2?: number | null;
  /** @nullable */
  aircraft?: number | null;
  /** @nullable */
  fighterJets?: number | null;
  /** @nullable */
  attackHelicopters?: number | null;
  /** @nullable */
  transportHelicopters?: number | null;
  /** @nullable */
  tanks?: number | null;
  /** @nullable */
  armoredVehicles?: number | null;
  /** @nullable */
  selfPropelledArtillery?: number | null;
  /** @nullable */
  rocketProjectors?: number | null;
  /** @nullable */
  navalVessels?: number | null;
  /** @nullable */
  submarines?: number | null;
  /** @nullable */
  destroyers?: number | null;
  /** @nullable */
  frigates?: number | null;
  /** @nullable */
  aircraftCarriers?: number | null;
  /** @nullable */
  nuclearWarheads?: number | null;
}