# lib/api-zod/src/generated/types/regionsMetadata.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */

export interface RegionsMetadata {
  continents: string[];
  regions: string[];
}