# lib/api-zod/src/generated/types/timelineEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Military Strength Comparator API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEntry {
  year: number;
  /** @nullable */
  defenseBudgetUsd?: number | null;
  /** @nullable */
  activePersonnel?: number | null;
  /** @nullable */
  gdpUsd?: number | null;
}